+++ CHANGELOG.md
@@ -1,4 +1,83 @@
-## [1.6.6](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.5...v1.6.6) (2026-04-12)
+
+## [1.6.6-dev.12](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.11...v1.6.6-dev.12) (2026-04-13)
+
+
+### Bug Fixes
+
+* ST-powersuiteai create new ([74aed56](https://github.com/michael-raquel/PowerIntake-WebApp/commit/74aed564613e1bca3e198ce7f4c8d9133d8ec4ed))
+
+## [1.6.6-dev.11](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.10...v1.6.6-dev.11) (2026-04-13)
+
+
+### Bug Fixes
+
+* Power Suite AI Logs Front End and minor changes ([ba614a2](https://github.com/michael-raquel/PowerIntake-WebApp/commit/ba614a2bf0d6ae5c81a73779e810e0dcadc2a6be))
+
+## [1.6.6-dev.10](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.9...v1.6.6-dev.10) (2026-04-13)
+
+
+### Reverts
+
+* Revert "Redirect to no-consent on 403" ([d442050](https://github.com/michael-raquel/PowerIntake-WebApp/commit/d44205093309a61e4b46365524feb95a2d5b571a))
+
+## [1.6.6-dev.9](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.8...v1.6.6-dev.9) (2026-04-13)
+
+
+### Bug Fixes
+
+* hid the notification on the settings ([2f384be](https://github.com/michael-raquel/PowerIntake-WebApp/commit/2f384be58b658e27b040774ce387ad5cd54beb01))
+
+## [1.6.6-dev.8](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.7...v1.6.6-dev.8) (2026-04-13)
+
+
+### Bug Fixes
+
+* my client table pagination row count and hide the notification ([6e210c6](https://github.com/michael-raquel/PowerIntake-WebApp/commit/6e210c6a359edef02d5d68720de5e16f7943029b))
+
+## [1.6.6-dev.7](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.6...v1.6.6-dev.7) (2026-04-13)
+
+
+### Bug Fixes
+
+* the v? error ([9dea823](https://github.com/michael-raquel/PowerIntake-WebApp/commit/9dea823e6ebf539d17e66c92f9f44ae8f486a1a3))
+
+## [1.6.6-dev.6](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.5...v1.6.6-dev.6) (2026-04-13)
+
+
+### Bug Fixes
+
+* updated filtering to fix the default filtering issue ([13b99a5](https://github.com/michael-raquel/PowerIntake-WebApp/commit/13b99a54fb5567b947959a3ed9d3383d63161815))
+
+## [1.6.6-dev.5](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.4...v1.6.6-dev.5) (2026-04-13)
+
+
+### Bug Fixes
+
+* remove socket ([538523d](https://github.com/michael-raquel/PowerIntake-WebApp/commit/538523d5b824820e76b0b5aef65637c5fd339523))
+
+## [1.6.6-dev.4](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.3...v1.6.6-dev.4) (2026-04-13)
+
+
+### Bug Fixes
+
+* notification socket ([74c488e](https://github.com/michael-raquel/PowerIntake-WebApp/commit/74c488e83922c82bceb46317874948c50a1d2b69))
+
+## [1.6.6-dev.3](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.2...v1.6.6-dev.3) (2026-04-13)
+
+
+### Bug Fixes
+
+* Fixed the button grouping ([663d76f](https://github.com/michael-raquel/PowerIntake-WebApp/commit/663d76fe7ff93688565b9176dcf60a1b00e3e3f0))
+
+## [1.6.6-dev.2](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.6-dev.1...v1.6.6-dev.2) (2026-04-13)
+
+
+### Bug Fixes
+
+* added a label asking if the ai was helpful ([de76562](https://github.com/michael-raquel/PowerIntake-WebApp/commit/de76562fce3f4e45958acdfa94015ae0e926cba6))
+
+## [1.6.6-dev.1](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.6.5...v1.6.6-dev.1) (2026-04-12)
+
 
 
 ### Bug Fixes